[upstream] unbreak using docker as client
Docker's newer clients popuates `cacheFrom` and `cacheTo` parameter
by default as empty array for all commands but buildah's design of
distributed cache expects this to be a repo not image hence parse
only the first populated repo and igore if empty array.

NB: Fixes building containers using the docker client with podman,
even when not using any cache-to/cache-from flag.