[backport] libpod: move oom_score_adj clamp to init
commit 8b4a79a744ac3fd176ca4dc0e3ae40f75159e090 introduced
oom_score_adj clamping when the container oom_score_adj value is lower
than the current one in a rootless environment.  Move the check to
init() time so it is performed every time the container starts and not
only when it is created.  It is more robust if the oom_score_adj value
is changed for the current user session.

Author: Giuseppe Scrivano <gscrivan@redhat.com>

Note: This, together with "clamp oom_score_adj if it is too low",
      unbreaks "docker compose up" with crun 1.9