integration-tests: Ignore os.Lchown failures
When this chown fails, it is more indicative of an issue with the
test environemnt rather than with the code under test.

I believe I've seen it in some interation on my laptop, but I am
haven't been able to determine the exact conditions. It does appear
to happen quite reliably on debci, though